More utility to utility module.
